Update Project2_Hardware.cpp
Modified parameters on blob detection, changed from background subtraction to absolute difference (works much better) Maybe up the upper tolerance? Also added inRange thresholding instead of single sided. Finished moving ROI and started working on a debug configuration to show the ROI and detected ball on the original image. Added bounds on the ROI so that it doesn't move out of frame (need to fix these, something is off)
(Need to move a lot of it into the if statement to speed up code execution.